Added ritual of imbuement

+++ src/main/java/com/leo/soulmancy/util/ShapeUtil.java
@@ -8,7 +8,7 @@ import net.minecraft.world.phys.shapes.VoxelShape;
 public class ShapeUtil {
 
 
-    public static VoxelShape shapeFromDimension(int x1, int y1, int z1, int x2, int y2, int z2){
+    public static VoxelShape shapeFromDimension(float x1, float y1, float z1, float x2, float y2, float z2){
         return Block.box(x1, y1, z1, x1 + x2, y1 + y2 , z1 + z2);
     }
